requirements.txt
fastmcp>=2.14.1
mcp>=1.25.0
langchain-mcp-adapters==0.1.12

pydantic==2.11.7
psutil
pyyaml
jinja2

